+++ app/Http/Controllers/StateNotesController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Note;
 use App\State;
-use Illuminate\Http\Request;
-use Mockery\Matcher\Not;
 
 class StateNotesController extends Controller
 {
@@ -17,24 +15,20 @@ class StateNotesController extends Controller
     public function store(State $state)
     {
         $this->validateWith(['body' => 'required']);
-
         return $state->addNote(request(['body']));
     }
 
     public function update(State $state, Note $note)
     {
-        $note = $state->notes()->findOrFail($note->id);
-
+        abort_unless($state->owns($note), 404);
         $this->validateWith(['body' => 'required']);
-
         $note->update(request(['body']));
-
         return $note;
     }
 
     public function destroy(State $state, Note $note)
     {
-        $note = $state->notes()->findOrFail($note->id);
+        abort_unless($state->owns($note), 404);
         $note->delete();
         return $note;
     }
